requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.0.0
